worker/requirements.txt
python-binance
pymongo
pika
python-dotenv